chore: organize routed MoE compiler modules
