+++ aten/models/gpt_oss/__init__.py
@@ -1,0 +1,1 @@
+"""GPT-OSS model semantics and reference utilities."""
